add discovery layer and implement kubernetes cluster discovery, also move discovery test code from datasource_store to discovery layer.
